requirements.txt
fastapi>=0.95.0  
uvicorn>=0.21.0  
sqlalchemy>=2.0.0  
python-jose[cryptography]  
passlib[bcrypt]  
redis>=4.5.0  
alembic>=1.10.0  
python-multipart  
python-dotenv  
pydantic-settings  